Solidus 2.3 upgrade
