+++ app/models/spree/line_item_decorator.rb
