+++ config/initializers/spree.rb
@@ -1,3 +1,3 @@
 Spree::RoleConfiguration.configure do |config|
-  config.assign_permissions :default, [Spree::Customizations::PermissionSets::DefaultCustomer]
+  config.assign_permissions :default, [SolidusCustomizations::PermissionSets::DefaultCustomer]
 end
