+++ lib/solidus_customizations/environment.rb
@@ -1,0 +1,13 @@
+module SolidusCustomizations
+  class Environment
+    include Core::EnvironmentExtension
+
+    add_class_set :source_configurations
+
+    attr_accessor :preferences
+
+    def initialize
+      @preferences = Spree::Customizations::AppConfiguration.new
+    end
+  end
+end
