+++ lib/solidus_customizations/factories.rb
