+++ lib/solidus_customizations/factories/customization_factory.rb
