--- lib/spree/permitted_attributes_decorator.rb
@@ -1,16 +1,0 @@
-Spree::PermittedAttributes.module_eval do
-
-  class_variable_set(:@@line_item_option_attributes,
-                     class_variable_get(:@@line_item_option_attributes).push(
-                         {
-                             customizations_attributes: [
-                                 :configuration_id,
-                                 :configuration_type,
-                                 :source_id,
-                                 :source_type,
-                                 :article_id,
-                                 :article_type
-                             ]
-                         }
-                     ))
-end
